backend/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==23.1.0
azure-core==1.26.4
azure-identity==1.12.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.2
frozenlist==1.3.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.17.0
httpx==0.24.0
hyperframe==6.0.1
idna==3.4
microsoft-kiota-abstractions==0.5.1
microsoft-kiota-authentication-azure==0.2.0
microsoft-kiota-http==0.4.1
microsoft-kiota-serialization-json==0.3.2
microsoft-kiota-serialization-text==0.2.0
msal==1.22.0
msal-extensions==1.0.0
msgraph-core==1.0.0a4
msgraph-sdk==1.0.0a12
multidict==6.0.4
portalocker==2.7.0
pycparser==2.21
PyJWT==2.6.0
python-dateutil==2.8.2
requests==2.30.0
six==1.16.0
sniffio==1.3.0
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==2.0.2
yarl==1.9.2
